+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Improve error messages
+
 # 0.3.1 - 2018-12-26
 
 * Update minimum derive_utils version to 0.5.1